requirements.txt
customtkinter
google-genai
fpdf
pandas
matplotlib